updated uart_write_hex for 64-bit numbers

+++ include/drivers/uart.h
@@ -7,7 +7,7 @@ void uart_init();
 void uart_tx(char ch);
 char uart_rx();
 void uart_write(char* str);
-void uart_write_hex(unsigned int x);
+void uart_write_hex(unsigned long long x);
 void uart_write_size(char* str, uint64_t size);
